Huge fix in `SlaveController` and `IntruderTileDispenser` instantiation. The fact is that there was a problem while extending those abstract classes. We didn't understand that we had to completely decouple object contrsuction and its initialization.
Controllers and Dispensers need some abstract methods to be called to be
fully initialized. The problem is that, while programming, it comes
natural to pass arguments in extending classes' constructors and use
them in abstract methods implementation. If these abstract methods are
called in super's constructor (and it is so...) it happens something
like (take `SlaveController` as example):

    ```
    public abstract class SlaveController {

        private TileDispenser dispenser;

        public SlaveController() {
            super();
            dispenser = getDispenser();
        }

        protected abstract TileDispenser getDispenser();

    }

    ...

    public class SlaveImpl extends SlaveController {
        private Object initArg;

        public SlaveImpl(Object initArg){
            super();
            // this super() call will invoke `getDispenser`
            // that will use `initArg` that, up to now,
            // it is null!!!!
            this.initArg = initArg;
        }

        @Override
        TileDispenser getDispenser(){
            return new DispenserImpl(initArg);
        }
    }
    ```

The solution is to decouple instantiation and initialization using an
`init` method.
The CONS are that we have to remember to `.init()` dispensers and
slavecontrollers after `new`; if not, NullPointerExceptions as fuck.

In addition, this commit reverts:

- commit 01075e73443c95e534944d9c970cdeb58306e078.
"Fix in ts passing in IntruderDispensers constructors."

- commit 1f531a2922fd0a62f363730752af05db9499475e.
"Every SlaveController have to take a TileSelector on instantiation"